replacing Mind with networkx
