+++ knowledge_graph/ontology_processing_utils.py
@@ -1,5 +1,6 @@
 import networkx as nx
 from networkx.readwrite import json_graph
+from knowledge_graph import app
 import os
