+++ process_new_ontology_file.py
@@ -2,8 +2,10 @@
 
 import os
 import argparse
+from knowledge_graph import app
 import knowledge_graph.make_network
 import knowledge_graph.make_graph
+from knowledge_graph.network_x_processor import network_x_processor
 
 
 def newest(path):
@@ -74,6 +76,15 @@ def main(args):
 
     # process the OWL ontology file
     processOntology(onto_path=onto_path, output_folder_path=output_folder_path)
+
+    try:
+        nx_processor = network_x_processor()
+    except:
+        raise FileNotFoundError("Pickle File Failed to Load as Config Variable")
+    try:
+        app.config["G"] = nx_processor.get_graph()
+    except:
+        raise ValidationError("Unable to Load Graph")
 
 
 if __name__ == "__main__":
